Extract RecipeListItem to its own component.
